+++ w4/p2/ConfirmOrder.h
@@ -15,7 +15,7 @@ namespace sdds{
 
     class ConfirmOrder{
         const sdds::Toy** m_orderArray{nullptr};
-        int m_numOfConfirms{};
+        size_t m_numOfConfirms{};
 
     public:
         ConfirmOrder(){};
